[Shaders] Make sure SampleCmp is declared as returning float

+++ sources/common/shaders/SiliconStudio.Shaders/Analysis/Hlsl/HlslDeclarations.h
@@ -47,8 +47,8 @@ class __Texture1D<T> {
 	float4 Sample(sampler_state S, float Location, int Offset);
 	float4 SampleBias(sampler_state S, float Location, float Bias);
 	float4 SampleBias(sampler_state S, float Location, float Bias, int Offset);
-	uint SampleCmp(sampler_state S, float Location, float Bias);
-	uint SampleCmp(sampler_state S, float Location, float Bias, int Offset);
+	float SampleCmp(sampler_state S, float Location, float Bias);
+	float SampleCmp(sampler_state S, float Location, float Bias, int Offset);
 	float4 SampleGrad(sampler_state S, float Location, float DDX, float DDY);
 	float4 SampleGrad(sampler_state S, float Location, float DDX, float DDY, int Offset);
 	float4 SampleLevel( sampler_state S, float Location, float LOD);
@@ -73,8 +73,8 @@ class __Texture1DArray<T> {
 	float4 Sample(sampler_state S, float2 Location, int Offset);
 	float4 SampleBias(sampler_state S, float2 Location, float Bias);
 	float4 SampleBias(sampler_state S, float2 Location, float Bias, int Offset);
-	uint SampleCmp(sampler_state S, float2 Location, float Bias);
-	uint SampleCmp(sampler_state S, float2 Location, float Bias, int Offset);
+	float SampleCmp(sampler_state S, float2 Location, float Bias);
+	float SampleCmp(sampler_state S, float2 Location, float Bias, int Offset);
 	float4 SampleGrad(sampler_state S, float2 Location, float DDX, float DDY);
 	float4 SampleGrad(sampler_state S, float2 Location, float DDX, float DDY, int Offset);
 	float4 SampleLevel( sampler_state S, float2 Location, float LOD);
@@ -101,8 +101,8 @@ class __Texture2D<T> {
 	float4 Sample(sampler_state S, float2 Location, int2 Offset);
 	float4 SampleBias(sampler_state S, float2 Location, float Bias);
 	float4 SampleBias(sampler_state S, float2 Location, float Bias, int2 Offset);
-	uint SampleCmp(sampler_state S, float2 Location, float Bias);
-	uint SampleCmp(sampler_state S, float2 Location, float Bias, int2 Offset);
+	float SampleCmp(sampler_state S, float2 Location, float Bias);
+	float SampleCmp(sampler_state S, float2 Location, float Bias, int2 Offset);
 	float4 SampleGrad(sampler_state S, float2 Location, float2 DDX, float2 DDY);
 	float4 SampleGrad(sampler_state S, float2 Location, float2 DDX, float2 DDY, int2 Offset);
 	float4 SampleLevel( sampler_state S, float2 Location, float LOD);
@@ -194,8 +194,8 @@ class __Texture2DArray<T> {
 	float4 Sample(sampler_state S, float3 Location, int2 Offset);
 	float4 SampleBias(sampler_state S, float3 Location, float Bias);
 	float4 SampleBias(sampler_state S, float3 Location, float Bias, int2 Offset);
-	uint SampleCmp(sampler_state S, float3 Location, float Bias);
-	uint SampleCmp(sampler_state S, float3 Location, float Bias, int2 Offset);
+	float SampleCmp(sampler_state S, float3 Location, float Bias);
+	float SampleCmp(sampler_state S, float3 Location, float Bias, int2 Offset);
 	float4 SampleGrad(sampler_state S, float3 Location, float2 DDX, float2 DDY);
 	float4 SampleGrad(sampler_state S, float3 Location, float2 DDX, float2 DDY, int2 Offset);
 	float4 SampleLevel( sampler_state S, float3 Location, float LOD);
@@ -286,8 +286,8 @@ class __Texture3D<T> {
 	float4 Sample(sampler_state S, float3 Location, int3 Offset);
 	float4 SampleBias(sampler_state S, float3 Location, float Bias);
 	float4 SampleBias(sampler_state S, float3 Location, float Bias, int3 Offset);
-	uint SampleCmp(sampler_state S, float3 Location, float Bias);
-	uint SampleCmp(sampler_state S, float3 Location, float Bias, int3 Offset);
+	float SampleCmp(sampler_state S, float3 Location, float Bias);
+	float SampleCmp(sampler_state S, float3 Location, float Bias, int3 Offset);
 	float4 SampleGrad(sampler_state S, float3 Location, float3 DDX, float3 DDY);
 	float4 SampleGrad(sampler_state S, float3 Location, float3 DDX, float3 DDY, int3 Offset);
 	float4 SampleLevel( sampler_state S, float3 Location, float LOD);
@@ -309,7 +309,7 @@ class __TextureCube<T> {
 	void GetDimensions( out float Width, out float Height);
 	float4 Sample(sampler_state S, float3 Location);
 	float4 SampleBias(sampler_state S, float3 Location, float Bias);
-	uint SampleCmp(sampler_state S, float3 Location, float Bias);
+	float SampleCmp(sampler_state S, float3 Location, float Bias);
 	float4 SampleGrad(sampler_state S, float3 Location, float3 DDX, float3 DDY);
 	float4 SampleLevel( sampler_state S, float3 Location, float LOD);
 };
@@ -324,8 +324,8 @@ class __TextureCubeArray<T> {
 	void GetDimensions( out float Width, out float Height, out float Elements);
 	float4 Sample(sampler_state S, float4 Location);
 	float4 SampleBias(sampler_state S, float4 Location, float Bias);
-	uint SampleCmp(sampler_state S, float4 Location, float Bias);
-	uint SampleCmp(sampler_state S, float4 Location, float Bias, int3 Offset);
+	float SampleCmp(sampler_state S, float4 Location, float Bias);
+	float SampleCmp(sampler_state S, float4 Location, float Bias, int3 Offset);
 	float4 SampleGrad(sampler_state S, float4 Location, float3 DDX, float3 DDY);
 	float4 SampleLevel( sampler_state S, float4 Location, float LOD);
 };
